# optimize `Exp` for even bases