--- a/uint256.go
+++ b/uint256.go
@@ -1182,25 +1182,16 @@ func (z *Int) Byte(n *Int) *Int {
 
 // Exp sets z = base**exponent mod 2**256, and returns z.
 func (z *Int) Exp(base, exponent *Int) *Int {
-	res := Int{1, 0, 0, 0}
-	multiplier := *base
-	expBitLen := exponent.BitLen()
-
-	curBit := 0
-	word := exponent[0]
-
-	if base[0]&1 == 0 {
-		if expBitLen > 8 {
-			return z.Clear()
-		}
-		for ; curBit < expBitLen; curBit++ {
-			if word&1 == 1 {
-				res.Mul(&res, &multiplier)
-			}
-			multiplier.squared()
-			word >>= 1
-		}
-		return z.Set(&res)
+	var (
+		res        = Int{1, 0, 0, 0}
+		multiplier = *base
+		expBitLen  = exponent.BitLen()
+		curBit     = 0
+		word       = exponent[0]
+		even       = base[0]&1 == 0
+	)
+	if even && expBitLen > 8 {
+		return z.Clear()
 	}
 
 	for ; curBit < expBitLen && curBit < 64; curBit++ {
@@ -1209,6 +1200,9 @@ func (z *Int) Exp(base, exponent *Int) *Int {
 		}
 		multiplier.squared()
 		word >>= 1
+	}
+	if even { // If the base was even, we are finished now
+		return z.Set(&res)
 	}
 
 	word = exponent[1]