@holiman 2024-12-06
Sorry for the delay. 
I have rewitten the benchmarks -- the 'even' checks were always hitting the early exit due to >8bit exponent. Now the even-tests are more fair, at exactly 8 bits exponent. 

I also simplified your implementation a bit, and cleaned up the declarations. PTAL